Refactor code

+++ js/reader.js
@@ -1,17 +1,17 @@
-const { dialog, BrowserWindow, app } = require("electron");
-const path = require("path");
-const util = require("util");
-const execFile = util.promisify(require("child_process").execFile);
-const fs = require("fs-extra");
+const { dialog, BrowserWindow, app } = require('electron');
+const path = require('path');
+const util = require('util');
+const execFile = util.promisify(require('child_process').execFile);
+const fs = require('fs-extra');
 
 let loadedFilePath = null;
 let images = [];
 
-const imageExt = [".png", ".jpg"];
+const imageExt = ['.png', '.jpg'];
 
-const resourcePath = app.isPackaged ? process.resourcesPath : ".";
+const resourcePath = app.isPackaged ? process.resourcesPath : '.';
 
-const getFiles = async (path = "./") => {
+const getFiles = async (path = './') => {
   const entries = await fs.readdir(path, { withFileTypes: true });
   const files = entries
     .filter((file) => !file.isDirectory())
@@ -25,36 +25,36 @@ const getFiles = async (path = "./") => {
 };
 
 const loadRcc = async (filePath) => {
-  const localPath = `${path.resolve(resourcePath, "rcc")}`;
+  const localPath = path.resolve(resourcePath, 'rcc');
 
   // clear previous images
   images = [];
 
   // delete res directory
-  try {
-    await fs.rmdir(path.join(localPath, "qresource"), { recursive: true });
-  } catch {}
+  if (fs.existsSync(path.join(localPath, 'qresource'))) {
+    await fs.rmdir(path.join(localPath, 'qresource'), { recursive: true });
+  }
 
-  await fs.copyFile(filePath, path.join(localPath, "res.rcc"));
+  await fs.copyFile(filePath, path.join(localPath, 'res.rcc'));
 
-  await execFile(path.join(localPath, "rcc.exe"), ["--reverse"], {
+  await execFile(path.join(localPath, 'rcc.exe'), ['--reverse'], {
     cwd: `${localPath}/`,
   });
 
   // get directory content
   const files = await getFiles(
-    path.join(localPath, "qresource", "res", "res.rcc")
+    path.join(localPath, 'qresource', 'res', 'res.rcc')
   );
 
   for (const file of files) {
     images.push({
       name: path.parse(file.name).name,
       path: path.relative(
-        path.join(localPath, "qresource", "res", "res.rcc"),
+        path.join(localPath, 'qresource', 'res', 'res.rcc'),
         file.path
       ),
       isImage: imageExt.includes(path.extname(file.path)),
-      data: Buffer.from(await fs.readFile(file.path, "binary"), "binary"),
+      data: Buffer.from(await fs.readFile(file.path, 'binary'), 'binary'),
     });
   }
 
@@ -62,29 +62,29 @@ const loadRcc = async (filePath) => {
   images.sort((a, b) => a.name.localeCompare(b.name));
 
   // cleanup
-  await fs.rmdir(path.join(localPath, "qresource"), { recursive: true });
-  await fs.rm(path.join(localPath, "res.rcc"));
+  await fs.rmdir(path.join(localPath, 'qresource'), { recursive: true });
+  await fs.rm(path.join(localPath, 'res.rcc'));
 
   loadedFilePath = filePath;
 
-  BrowserWindow.getAllWindows()[0].webContents.send("populate-list", images);
+  BrowserWindow.getAllWindows()[0].webContents.send('populate-list', images);
 };
 
 const extractToPng = async (directoryPath) => {
   if (!images.length) {
-    dialog.showErrorBox("Error", "Nothing to extract.");
+    dialog.showErrorBox('Error', 'Nothing to extract.');
     return;
   }
 
   for (const image of images) {
     if (image.isImage) {
-      await fs.outputFile(`${directoryPath}/${image.path}`, image.data);
+      await fs.outputFile(path.join(directoryPath.image.path), image.data);
     }
   }
 
   dialog.showMessageBox(null, {
     message: `Png images extracted successfully. Extracted ${images.length} images.`,
-    type: "info",
+    type: 'info',
   });
 };
 
@@ -93,49 +93,49 @@ const saveRcc = async (filePath = loadedFilePath) => {
     return;
   }
 
-  const localPath = path.resolve(resourcePath, "rcc");
+  const localPath = path.resolve(resourcePath, 'rcc');
 
   // create .qrc file
-  let data = `<!DOCTYPE RCC><RCC version="1.0">\n<qresource>\n`;
+  let data = '<!DOCTYPE RCC><RCC version="1.0">\n<qresource>\n';
 
   for (const image of images) {
     data += `<file>${image.path}</file>\n`;
   }
 
-  data += `</qresource>\n</RCC>`;
+  data += '</qresource>\n</RCC>';
 
-  await fs.outputFile(path.join(localPath, "res", "res.qrc"), data);
+  await fs.outputFile(path.join(localPath, 'res', 'res.qrc'), data);
 
   // dump images
   for (const image of images) {
-    await fs.outputFile(path.join(localPath, "res", image.path), image.data);
+    await fs.outputFile(path.join(localPath, 'res', image.path), image.data);
   }
 
   await execFile(
-    path.join(localPath, "rcc.exe"),
+    path.join(localPath, 'rcc.exe'),
     [
-      "--format-version",
-      "1",
-      "--binary",
-      `./res/res.qrc`,
-      "-o",
-      "./res/res_output.rcc",
+      '--format-version',
+      '1',
+      '--binary',
+      './res/res.qrc',
+      '-o',
+      './res/res_output.rcc',
     ],
     {
       cwd: `${localPath}/`,
     }
   );
 
-  await fs.move("./rcc/res/res_output.rcc", filePath, { overwrite: true });
+  await fs.move('./rcc/res/res_output.rcc', filePath, { overwrite: true });
 
   // cleanup
-  await fs.rmdir(path.join(localPath, "res"), {
+  await fs.rmdir(path.join(localPath, 'res'), {
     recursive: true,
   });
 
   dialog.showMessageBox(null, {
-    message: `Rcc saved successfully.`,
-    type: "info",
+    message: 'Rcc saved successfully.',
+    type: 'info',
   });
 };
 
@@ -145,13 +145,11 @@ const replaceImage = async (index, filePath) => {
     return null;
   }
 
-  image.data = Buffer.from(await fs.readFile(filePath, "binary"), "binary");
+  image.data = Buffer.from(await fs.readFile(filePath, 'binary'), 'binary');
   return image.data;
 };
 
-const getImageByIndex = (index) => {
-  return images[index]?.data;
-};
+const getImageByIndex = (index) => images[index]?.data;
 
 module.exports = {
   loadRcc,
